+++ src/linkers/node-modules.ts
@@ -1,34 +1,51 @@
 import { Project, Package, structUtils } from "@yarnpkg/core";
 import { parseSyml } from "@yarnpkg/parsers";
-import { readFileSync } from "fs";
-import { npath, ppath, Filename } from "@yarnpkg/fslib";
+import { xfs, ppath, PortablePath, Filename } from "@yarnpkg/fslib";
+import { ManifestWithLicenseInfo } from ".";
 
-export const getPackageManifest = (project: Project, pkg: Package) => {
-  makeYarnState(project);
+/**
+ * Get package manifest with `node-modules` linker for a given Yarn project and package
+ *
+ * @param {Project} project - Yarn project
+ * @param {Package} pkg - Yarn package
+ * @returns {Promise<ManifestWithLicenseInfo | null>} Package manifest
+ */
+export const getPackageManifest = async (
+  project: Project,
+  pkg: Package
+): Promise<ManifestWithLicenseInfo | null> => {
+  await makeYarnState(project);
 
   const locator = structUtils.convertPackageToLocator(pkg);
   const entry = yarnState[structUtils.stringifyLocator(locator)];
   if (!entry) return null;
 
   const location = entry.locations[0];
-  const portablePath = location
+  const relativePath = location
     ? ppath.join(location, Filename.manifest)
     : Filename.manifest;
-  const nativePath = npath.fromPortablePath(portablePath);
-  const packageJson = readFileSync(nativePath).toString();
+  const portablePath = ppath.join(project.cwd, relativePath);
+  const packageJson = await xfs.readFilePromise(portablePath, "utf8");
 
   return JSON.parse(packageJson);
 };
 
-let yarnState;
-const makeYarnState = (project: Project) => {
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+let yarnState: any;
+
+/**
+ * Cache Yarn state from `yarn-state.yml`, if it has not already been cached
+ *
+ * @param {Project} project - Yarn project
+ * @returns {Promise<void>}
+ */
+const makeYarnState = async (project: Project): Promise<void> => {
   if (!yarnState) {
     const portablePath = ppath.join(
-      project.configuration.projectCwd,
+      project.configuration.projectCwd as PortablePath,
       Filename.nodeModules,
       ".yarn-state.yml" as Filename
     );
-    const nativePath = npath.fromPortablePath(portablePath);
-    yarnState = parseSyml(readFileSync(nativePath).toString());
+    yarnState = parseSyml(await xfs.readFilePromise(portablePath, "utf8"));
   }
 };