+++ src/linkers/pnp.ts
@@ -1,3 +1,6 @@
+/* istanbul ignore file */
+// this file is covered by CLI tests
+
 import { getPnpPath } from "@yarnpkg/plugin-pnp";
 import { Package, Project, structUtils } from "@yarnpkg/core";
 import {
@@ -8,8 +11,19 @@ import {
   ppath,
 } from "@yarnpkg/fslib";
 import { getLibzipSync } from "@yarnpkg/libzip";
+import { ManifestWithLicenseInfo } from ".";
 
-export const getPackageManifest = (project: Project, pkg: Package) => {
+/**
+ * Get package manifest with `pnp` linker for a given Yarn project and package
+ *
+ * @param {Project} project - Yarn project
+ * @param {Package} pkg - Yarn package
+ * @returns {Promise<ManifestWithLicenseInfo | null>} Package manifest
+ */
+export const getPackageManifest = async (
+  project: Project,
+  pkg: Package
+): Promise<ManifestWithLicenseInfo | null> => {
   makePnPApi(project);
 
   const locator = structUtils.convertPackageToLocator(pkg);
@@ -19,27 +33,37 @@ export const getPackageManifest = (project: Project, pkg: Package) => {
   };
 
   const packageInformation = pnpApi.getPackageInformation(pnpLocator);
-  if (!packageInformation) return;
+  if (!packageInformation) return null;
 
   const { packageLocation } = packageInformation;
   const portablePath: PortablePath = ppath.join(
     packageLocation,
     Filename.manifest
   );
-  const packageJson = fs.readFileSync(portablePath).toString();
+  const packageJson = await fs.readFilePromise(portablePath, "utf8");
 
   return JSON.parse(packageJson);
 };
 
-let pnpApi;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+let pnpApi: any;
+
+/**
+ * Cache PnP API from `.pnp.js` (or similar), if it has not already been cached
+ *
+ * @param {Project} project - Yarn project
+ * @returns {void}
+ */
 const makePnPApi = (project: Project) => {
   if (!pnpApi) {
     // use `eval` so webpack leaves this alone
-    // tslint:disable-next-line:no-eval
     pnpApi = eval("module.require")(getPnpPath(project).main);
   }
 };
 
+/**
+ * Instantiate the virtual file system for reading package manifests in PnP
+ */
 const fs = new VirtualFS({
   baseFs: new ZipOpenFS({
     libzip: getLibzipSync(),